refactor build script
