+++ CHANGELOG.md
@@ -3,7 +3,10 @@
 
 ## [0.16.0] - Unreleased
 ### Changes
-- [BREAKING] Remove image conversion code, reducing deps & can be delegated to the image crate.
+- [BREAKING] Remove image conversion code.
+- Refactor build script.
+- Enable the fltk-bundled feature on a system path using the CFLTK_BUNDLE_DIR env variable.
+- Enable the fltk-bundled feature on a user-defined url using CFLTK_BUNDLE_URL env var.
 - Add app::set_font_size() to set the global font size of the app.
 
 ## [0.15.15] - 2021-03-21
